* Changed some stuff.
When a player drops from the server, it sends a ClientDroppedEvent to Host.
The host will send a PlayerKickEvent to game. Game will delete Players and kick the client.
When kick is called, the client channel will notify the Host with a ClientDisconnectedEvent. Host will then delete the ClientData and Channel.